@@ -18,38 +18,38 @@ import pl.piomin.services.protobuf.customer.model.CustomerProto.Customers;
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
 public class CustomerApplicationTest {
 
-	protected Logger logger = Logger.getLogger(CustomerApplicationTest.class.getName());
-	
-	@Autowired
-	TestRestTemplate template;
-	
-//	@Test
-	public void testFindById() {
-		Customer c = this.template.getForObject("/customers/{id}", Customer.class, 1);
-		logger.info("Customer[\n" + c + "]");
-	}
-	
-	@Test
-	public void testFindByPesel() {
-		Customer c = this.template.getForObject("/customers/pesel/{pesel}", Customer.class, "12346");
-		logger.info("Customer[\n" + c + "]");
-	}
-	
-	@Test
-	public void testFindAll() {
-		Customers c = this.template.getForObject("/customers", Customers.class);
-		logger.info("Customers[\n" + c + "]");
-	}
-	
+    protected Logger logger = Logger.getLogger(CustomerApplicationTest.class.getName());
 
-	@TestConfiguration
-	static class Config {
+    @Autowired
+    TestRestTemplate template;
 
-		@Bean
-		public RestTemplateBuilder restTemplateBuilder() {
-			return new RestTemplateBuilder().additionalMessageConverters(new ProtobufHttpMessageConverter());
-		}
+    //	@Test
+    public void testFindById() {
+        Customer c = this.template.getForObject("/customers/{id}", Customer.class, 1);
+        logger.info("Customer[\n" + c + "]");
+    }
 
-	}
-	
+    @Test
+    public void testFindByPesel() {
+        Customer c = this.template.getForObject("/customers/pesel/{pesel}", Customer.class, "12346");
+        logger.info("Customer[\n" + c + "]");
+    }
+
+    @Test
+    public void testFindAll() {
+        Customers c = this.template.getForObject("/customers", Customers.class);
+        logger.info("Customers[\n" + c + "]");
+    }
+
+
+    @TestConfiguration
+    static class Config {
+
+        @Bean
+        public RestTemplateBuilder restTemplateBuilder() {
+            return new RestTemplateBuilder().additionalMessageConverters(new ProtobufHttpMessageConverter());
+        }
+
+    }
+
 }
